Work further on memory instructions

+++ instr-encoding/src/lib.rs
@@ -520,7 +520,7 @@ macro_rules! field_type {
     (csr)                 => { CsrIndex };
     (uimm: csr)           => { u8 };
     (imm: itype_unsigned) => { u32 };
-    (imm: itype_signed)   => { i32 };
+    (imm: itype_signed)   => { i16 };
     (imm: stype)          => { i16 };
     (imm: btype)          => { i16 };
     (imm: jtype)          => { i32 };
@@ -646,7 +646,7 @@ macro_rules! field_decode {
     ($bs:expr, csr)                 => { CsrIndex(($bs >> 20) as u16) };
     ($bs:expr, uimm: csr)           => { (($bs >> 15) & 0x1F) as u8 };
     ($bs:expr, imm: itype_unsigned) => { ((($bs & 0xFFF0_0000) as i32) >> 20) as u32 };
-    ($bs:expr, imm: itype_signed)   => { ((($bs & 0xFFF0_0000) as i32) >> 20) as i32 };
+    ($bs:expr, imm: itype_signed)   => { ((($bs & 0xFFF0_0000) as i32) >> 20) as i16 };
     ($bs:expr, imm: stype)          => { ((($bs & 0xFE00_0000) as i32) >> 20) as i16 | (($bs >> 7) & 0x1F) as i16 };
     ($bs:expr, imm: btype)          => {{
             // Not straight forward because we need to sign extend
